web_app_doc_spl/requirements.txt
flask==2.3.3
flask-cors==4.0.0
protobuf==4.24.4
requests==2.31.0
urllib3==2.0.7
grpcio==1.59.0
grpcio-status==1.59.0
gunicorn
gevent
pydantic==2.11.3
pydantic-settings==2.8.1
tabulate==0.9.0
google-auth==2.36.0
requests==2.31.0
google-adk
google-cloud-aiplatform
llama-index==0.12
absl-py==2.1.0
cloudpickle==3.0.0
python-dotenv==1.1.0